refactor: move fundamental databse interactions to `services` (#131)
