+++ pages/api/atlases.ts
@@ -4,11 +4,11 @@ import {
 } from "../../app/apis/catalog/hca-atlas-tracker/common/entities";
 import { dbAtlasToApiAtlas } from "../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../app/common/entities";
+import { query } from "../../app/services/database";
 import {
   getUserRoleFromAuthorization,
   handler,
   method,
-  query,
 } from "../../app/utils/api-handler";
 
 /**
